Website Update
Added an event listener to run an anonymous function when the page loads that creates onblur event handlers for the input boxes with the IDs of fnBox, lnBox, groupBox, mail- Box, phoneBox, and banquetBox, running the calcCart() function in response to each event. Also created the sessionTest() function that provides a validation test for the conference session selection list, and started the calcCart() function that calculates the cost of the registration and saves data in session storage.