Website Update
Finished the calcCart function by creating an if statement that checks if the selected index of the sessionBox selection list is not equal to -1, and if its not it stores the text of the selected option in the session storage variable confSession and the value of the selected option in the confSessionCost session storage variable; otherwise stores an empty text string and the value 0 respectively. Also checks If the user has clicked the mediaCB check box, and if the user has, stores the values “yes” and 115 in the confPack and confPackcost session storage variables; otherwise store the values “no” and 0.